Modifications per recommended changes
* moved classes to org.tensorflow.op package
* BUILD changes to keep op package in a separate target
* Updated ci_build to also build new target
* Replaced static factory method for Scope with public raw constructor
* Replaced Builders for NameScope and Scope with raw constructors
* Addressed JavaDoc nits
* Added comments pointing to reasons for operator name restrictions
* Merged NameScopeTest into ScopeTest
* renamed scope.withOpName -> scope.withName
